PN-3044 - factorize render logic of AppStatus page in pn-commons - working, lacks test of the new, generic, component, and use in pf-webapp and pa-webapp
