+++ packages/pn-commons/src/types/index.ts
@@ -1,6 +1,7 @@
 // types
 import { PaginationData } from '../components/Pagination/types';
 import { DatePickerTypes } from '../components/CustomDatePicker';
+import { AppStatusData } from './AppStatus';
 import { IAppMessage } from './AppMessage';
 import { NotificationStatus } from './NotificationStatus';
 import { SideMenuItem } from './SideMenuItem';
@@ -59,6 +60,7 @@ export {
 };
 export type {
   AnalogWorkflowDetails,
+  AppStatusData,
   BasicUser,
   CardAction,
   CardElement,
